+++ ipyvizzu.py
@@ -8,6 +8,7 @@ import typing
 import uuid
 import enum
 from textwrap import indent, dedent
+import numpy as np
 
 from IPython.display import display_html
 
@@ -146,6 +147,13 @@ class PlainAnimation(dict, Animation):
         return self
 
 
+class InferType(enum.Enum):
+
+    DIMENSION = "dimension"
+    MEASURE = "measure"
+    AUTO = None
+
+
 class Data(dict, Animation):
     """
     Vizzu data with the required keys: records, series, dimensions or measures.
@@ -184,6 +192,38 @@ class Data(dict, Animation):
 
     def add_measure(self, name, values=None, **kwargs):
         self._add_named_value("measures", name, values, **kwargs)
+
+    def add_data_frame(
+        self,
+        data_frame,
+        infer_types=None,
+        default_measure_value=0,
+        default_dimension_value="",
+    ):
+        if infer_types is None:
+            infer_types = {}
+        for name in data_frame.columns:
+            infer_type = InferType(infer_types.get(name, InferType.AUTO))
+            if infer_type == InferType.AUTO:
+                if isinstance(data_frame[name].values[0], (np.float64, np.int64)):
+                    infer_type = InferType.MEASURE
+                else:
+                    infer_type = InferType.DIMENSION
+
+            values = []
+            if infer_type == InferType.MEASURE:
+                values = [
+                    float(i)
+                    for i in data_frame[name].fillna(default_measure_value).values
+                ]
+            else:
+                values = list(data_frame[name].fillna(default_dimension_value).values)
+
+            self.add_series(
+                name,
+                values,
+                type=infer_type.value,
+            )
 
     def _add_named_value(self, dest, name, values=None, **kwargs):
         value = {"name": name, **kwargs}