+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Added
 
+- Added `Data.add_data_frame()` function to support `pandas.DataFrame`.
 - Added `scroll_into_view` chart property.
 
 ## [0.6.0] - 2022-03-17